add python tests
