+++ .gitignore
@@ -15,6 +15,7 @@ build
 flatnav_python/flatnav.egg-info/
 flatnav_python/poetry.lock
 flatnav_python/dist
+flatnav_python/__pycache__
 
 
 # other files
