+++ flatnav_python/pyproject.toml
@@ -21,6 +21,9 @@ setuptools = "68.2.2"
 black = "^23.11.0"
 pytest = "^7.4.3"
 numpy = "^1.26.2"
+h5py = "^3.10.0"
+requests = "^2.31.0"
+
 
 
 [build-system]
