+++ flatnav_python/setup.py
@@ -1,21 +1,11 @@
-# import toml
 import os
-
-# Available at setup time due to pyproject.toml
 from pybind11.setup_helpers import Pybind11Extension, build_ext
 from setuptools import setup
-
-# def parse_version_from_pyproject() -> str:
-#     with open("pyproject.toml") as f:
-#         pyproject = toml.load(f)
-#         return pyproject["tool"]["poetry"]["version"]
-
-#     raise RuntimeError("Unable to find version string.")
 
 __version__ = "0.0.1"
 
 CURRENT_DIR = os.getcwd()
-SOURCE_PATH = os.path.join(CURRENT_DIR, "bindings.cpp")
+SOURCE_PATH = os.path.join(CURRENT_DIR, "python_bindings.cpp")
 
 
 ext_modules = [
@@ -28,9 +18,16 @@ ext_modules = [
             os.path.join(CURRENT_DIR, ".."),
             os.path.join(CURRENT_DIR, "..", "external", "cereal", "include"),
         ],
-        # Ignoring the `Wno-sign-compare` which warns you when you compare int with something like
-        # uint64_t.
-        extra_compile_args=["-Wno-sign-compare", "-fopenmp"],
+        extra_compile_args=[
+            "-fopenmp",  # Enable OpenMP
+            "-Ofast",  # Use the fastest optimization
+            "-fpic",  # Position-independent code
+            "-w",  # Suppress all warnings (note: this overrides -Wall)
+            "-ffast-math",  # Enable fast math optimizations
+            "-funroll-loops",  # Unroll loops
+            "-ftree-vectorize",  # Vectorize where possible
+        ],
+        extra_link_args=["-fopenmp"],  # Link OpenMP when linking the extension
     )
 ]
 
@@ -44,9 +41,6 @@ setup(
     description="Graph kNN with reordering.",
     long_description="",
     ext_modules=ext_modules,
-    # extras_require={"test": "pytest"},
-    # Currently, build_ext only provides an optional "highest supported C++
-    # level" feature, but in the future it may provide more features.
     cmdclass={"build_ext": build_ext},
     zip_safe=False,
     python_requires=">=3.7",
